rotation_calibration: extract reusable science module
Factor the pure calibration logic out of the REPL CLI into a shared
device-layer module so the forthcoming web UI can consume the same
solver, refraction model, prior-inspection, and JSON writer without
duplication.

New `device/rotation_calibration.py` owns:
- `Sighting`, `RotationSolution`, `PriorInfo` dataclasses
- `KEEP_MAX_AGE_S` / `KEEP_MAX_DISTANCE_M` constants
- `terrestrial_refraction_deg`, `predict_mount_azel`, `solve_rotation`,
  `write_calibration`
- `parse_calibrated_at`, `inspect_prior`, `decide_clear_or_keep`
  (previously private `_` prefixes — now public for cross-module use)

`scripts/trajectory/calibrate_rotation.py` re-imports these at the
top of the module and keeps only its REPL glue: argparse, input()
loops (`_handle_clear_or_keep`, `_altitude_menu`, `_nudge_loop`),
menu rendering, and the `_main_live` / `_main_replay` flow. An
`__all__` block re-exports the science names so existing test
imports from the script path keep working.

No behaviour change: CLI `--help` renders the same; unit tests
updated to import the dataclasses from the new module path; all 56
calibrate_rotation + faa_dof + target_frame tests pass; ruff clean.
